Updated typescript version
