+++ TypeScript/babystep.html
@@ -3,6 +3,9 @@
         
     </head>
     <body>
-        <script src="./babystep/babystep.js"></script>
+<script type="module">
+import { command } from "./src/babystep.js";
+window.command = command;
+</script>
     </body>
 </html>
