+++ TypeScript/src/babystep.ts
@@ -1,24 +1,32 @@
-"use strict";
-var BackgroundColorNeutral = "#ffffff";
-var BackgroundColorFailed = "#ffcccc";
-var BackgroundColorPassed = "#ccffcc";
-var SecondsInCycle = 120;
-var _timerRunning;
-var _currentCycleStartTime;
-var _lastRemainingTime;
-var _bodyBackgroundColor = BackgroundColorNeutral;
-var _threadTimer;
+const BackgroundColorNeutral: string = "#ffffff";
+const BackgroundColorFailed: string = "#ffcccc";
+const BackgroundColorPassed: string = "#ccffcc";
+const SecondsInCycle: number = 120;
+
+
+let _timerRunning: boolean;
+let _currentCycleStartTime: number;
+let _lastRemainingTime: string;
+let _bodyBackgroundColor: string = BackgroundColorNeutral;
+let _threadTimer: NodeJS.Timeout;
+
+
 document.body.innerHTML = CreateTimerHtml(getRemainingTimeCaption(0), BackgroundColorNeutral, false);
-function command(arg) {
-    var args = { Url: { AbsoluteUri: "command://" + arg + "/" } };
+
+
+export function command(arg: string): void {
+    let args = { Url: { AbsoluteUri: `command://${arg}/` } }
     console.log('called', arg, args.Url.AbsoluteUri);
     if (args.Url.AbsoluteUri == "command://start/") {
         document.body.innerHTML = CreateTimerHtml(getRemainingTimeCaption(0), BackgroundColorNeutral, true);
+
         _timerRunning = true;
         _currentCycleStartTime = Date.now();
-        _threadTimer = setInterval(function () {
+
+        _threadTimer = setInterval(function() {
             if (_timerRunning) {
-                var elapsedTime = Date.now() - _currentCycleStartTime;
+                let elapsedTime: number = Date.now() - _currentCycleStartTime;
+
                 if (elapsedTime >= SecondsInCycle * 1000 + 980) {
                     _currentCycleStartTime = Date.now();
                     elapsedTime = Date.now() - _currentCycleStartTime;
@@ -26,8 +34,12 @@ function command(arg) {
                 if (elapsedTime >= 5000 && elapsedTime < 6000 && _bodyBackgroundColor != BackgroundColorNeutral) {
                     _bodyBackgroundColor = BackgroundColorNeutral;
                 }
-                var remainingTime = getRemainingTimeCaption(elapsedTime);
+
+                let remainingTime: string = getRemainingTimeCaption(elapsedTime);
+             
+                
                 if (_lastRemainingTime !== remainingTime) {
+                   
                     if (remainingTime == "00:10") {
                         playSound("2166__suburban-grilla__bowl-struck.wav");
                     }
@@ -35,6 +47,7 @@ function command(arg) {
                         playSound("32304__acclivity__shipsbell.wav");
                         _bodyBackgroundColor = BackgroundColorFailed;
                     }
+
                     document.body.innerHTML = CreateTimerHtml(remainingTime, _bodyBackgroundColor, true);
                     _lastRemainingTime = remainingTime;
                 }
@@ -43,8 +56,9 @@ function command(arg) {
     }
     else if (args.Url.AbsoluteUri == "command://stop/") {
         _timerRunning = false;
-        clearInterval(_threadTimer);
+        clearInterval(_threadTimer)
         document.body.innerHTML = CreateTimerHtml(getRemainingTimeCaption(0), BackgroundColorNeutral, false);
+
     }
     else if (args.Url.AbsoluteUri == "command://reset/") {
         _currentCycleStartTime = Date.now();
@@ -52,24 +66,26 @@ function command(arg) {
     }
     else if (args.Url.AbsoluteUri == "command://quit/") {
         document.body.innerHTML = "";
-        clearInterval(_threadTimer);
+        clearInterval(_threadTimer)
     }
+
+};
+
+
+export function getRemainingTimeCaption(elapsedTime: number): string {
+
+    let remainingTime: Date = new Date((SecondsInCycle * 1000) - elapsedTime);
+    var minute: string | number = remainingTime.getMinutes();
+    var second: string | number = remainingTime.getSeconds();
+    if (minute < 10) { minute = '0' + minute; }
+    if (second < 10) { second = '0' + second; }
+
+    return '' + minute + ':' + second
 }
-;
-function getRemainingTimeCaption(elapsedTime) {
-    var remainingTime = new Date((SecondsInCycle * 1000) - elapsedTime);
-    var minute = remainingTime.getMinutes();
-    var second = remainingTime.getSeconds();
-    if (minute < 10) {
-        minute = '0' + minute;
-    }
-    if (second < 10) {
-        second = '0' + second;
-    }
-    return '' + minute + ':' + second;
-}
-function CreateTimerHtml(timerText, bodyColor, running) {
-    var timerHtml = "<div style=\"border: 3px solid #555555; background: " + bodyColor +
+
+export function CreateTimerHtml(timerText: string, bodyColor: string, running: boolean): string {
+
+    let timerHtml: string = "<div style=\"border: 3px solid #555555; background: " + bodyColor +
         "; margin: 0; padding: 0;\">" +
         "<h1 style=\"text-align: center; font-size: 30px; color: #333333;\">" + timerText +
         "</h1>" +
@@ -82,12 +98,14 @@ function CreateTimerHtml(timerText, bodyColor, running) {
         timerHtml += "<a style=\"color: #555555;\" href=\"javascript:command('start');\">Start</a> ";
     }
     timerHtml += "<a style=\"color: #555555;\" href=\"javascript:command('quit');\">Quit</a> ";
-    timerHtml += "</div></div>";
+    timerHtml += "</div></div>"
     return timerHtml;
+
 }
-function playSound(url) {
-    var audio = new Audio();
-    audio.src = "./babystep/sounds/" + url;
+
+function playSound(url: string): void {
+    let audio = new Audio();
+    audio.src = `./src/sounds/${url}`;
     console.log(audio.src);
     audio.load();
     audio.play();
